Extract tab system into reusable TabPanel component
Refactored the DevSessionUI to improve separation of concerns:

- Created new TabPanel component for reusable tab functionality
- Moved tab state management and input handling to TabPanel
- Kept global concerns (Ctrl+C, error handling) in DevSessionUI
- Cleaned up component APIs by only exporting necessary types
- Improved code organization and maintainability

The TabPanel component now handles:
- Tab state management (activeTab)
- Tab navigation input handling
- Tab rendering with box drawing characters
- Tab-specific shortcuts

DevSessionUI maintains responsibility for:
- Global input handling (Ctrl+C)
- Application lifecycle management
- Error state management
- Business logic coordination
