+++ packages/app/src/cli/services/dev/ui/components/TabPanel.tsx
@@ -1,0 +1,123 @@
+import React, {useState} from 'react'
+import {Box, Text, useInput, useStdin} from '@shopify/cli-kit/node/ink'
+
+export interface Tab {
+  label: string
+  content?: React.ReactNode
+  shortcuts?: TabShortcut[]
+  action?: () => Promise<void>
+}
+
+interface TabShortcut {
+  key: string
+  condition?: () => boolean
+  action: () => Promise<void>
+}
+
+interface TabDisplay extends Tab {
+  inputKey: string
+  header: string
+}
+
+interface TabPanelProps {
+  tabs: {[key: string]: Tab}
+  initialActiveTab: string
+}
+
+export const TabPanel: React.FunctionComponent<TabPanelProps> = ({tabs, initialActiveTab}) => {
+  const {isRawModeSupported: canUseShortcuts} = useStdin()
+  const [activeTab, setActiveTab] = useState<string>(initialActiveTab)
+
+  useInput(
+    (input) => {
+      const onInput = async () => {
+        // First check if input matches any tab key
+        const matchingTab = tabs[input]
+        if (matchingTab) {
+          if (matchingTab.action) {
+            await matchingTab.action()
+          } else {
+            setActiveTab(input)
+          }
+          return
+        }
+
+        // Then check if input matches any shortcut key for the current active tab
+        const currentTab = tabs[activeTab]
+        if (currentTab?.shortcuts) {
+          const matchingShortcut = currentTab.shortcuts.find((shortcut) => shortcut.key === input)
+          if (matchingShortcut) {
+            // Check condition if it exists
+            if (!matchingShortcut.condition || matchingShortcut.condition()) {
+              await matchingShortcut.action()
+            }
+          }
+        }
+      }
+
+      // eslint-disable-next-line @typescript-eslint/no-floating-promises
+      onInput()
+    },
+    {isActive: Boolean(canUseShortcuts)},
+  )
+
+  const tabsArray: TabDisplay[] = Object.entries(tabs).map(([key, tab]) => {
+    return {
+      ...tab,
+      inputKey: key,
+      header: ` (${key}) ${tab.label} `,
+    }
+  })
+
+  // We need to subtract the number of tabs from the total length to account for the borders between and after the tabs
+  const tabHeaderLength = tabsArray.reduce((acc, tab) => acc + (tab.header?.length ?? 0), 0) + tabsArray.length + 1
+
+  return (
+    <>
+      {/* Top border with connected tab boxes */}
+      <Text>
+        {tabsArray
+          .map((tab, index) => {
+            return `${index === 0 ? '╒' : '╤'}${'═'.repeat(tab.header?.length ?? 0)}`
+          })
+          .join('')}
+        {'╤'}
+        {'═'.repeat(Math.max(0, (process.stdout.columns || 100) - tabHeaderLength - 2))}
+        {'╕'}
+      </Text>
+      {/* Tab content row */}
+      <Text>
+        {tabsArray.map((tab) => {
+          return (
+            <React.Fragment key={tab.inputKey}>
+              {'│'}
+              <Text
+                bold={activeTab === tab.inputKey}
+                color={activeTab === tab.inputKey ? 'cyan' : 'white'}
+                backgroundColor={activeTab === tab.inputKey ? 'gray' : 'transparent'}
+              >
+                {tab.header}
+              </Text>
+            </React.Fragment>
+          )
+        })}
+        {`│${' '.repeat(Math.max(0, (process.stdout.columns || 100) - tabHeaderLength - 2))}│`}
+      </Text>
+      {/* Bottom border connecting tabs */}
+      <Text>
+        {tabsArray
+          .map((tab, index) => {
+            return `${index === 0 ? '└' : '┴'}${'─'.repeat(tab.header?.length ?? 0)}`
+          })
+          .join('')}
+        {'┴'}
+        {'─'.repeat(Math.max(0, (process.stdout.columns || 100) - tabHeaderLength - 2))}
+        {'┘'}
+      </Text>
+      {/* Tab Content Area */}
+      <Box flexDirection="column" marginLeft={1} marginRight={1}>
+        {tabs[activeTab]?.content}
+      </Box>
+    </>
+  )
+}
